+++ src/main/java/big/census/big_smallville_census_api/responses/AvgIncomeResponse.java
@@ -1,10 +1,12 @@
 package big.census.big_smallville_census_api.responses;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
 @AllArgsConstructor
 public class AvgIncomeResponse {
-    private Double avgIncome;
+    private BigDecimal avgIncome;
 }
