+++ src/main/java/big/census/big_smallville_census_api/services/BusinessService.java
@@ -1,6 +1,8 @@
 package big.census.big_smallville_census_api.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
@@ -14,7 +16,7 @@ public class BusinessService {
     @Autowired
     private BusinessRepository businessRepository;
 
-    public double getAverageIncome(String businessName) {
+    public BigDecimal getAverageIncome(String businessName) {
         return businessRepository.avgEmployeeIncome(businessName);
     }
